backend/apis/base_de_datos/requirements.txt
# Core FastAPI dependencies
fastapi>=0.100.0,<0.105.0
uvicorn[standard]>=0.20.0,<0.25.0
pydantic>=2.0.0,<3.0.0
pydantic[email]>=2.0.0,<3.0.0
python-multipart>=0.0.6

# Database
SQLAlchemy>=2.0.0,<2.1.0
psycopg2-binary>=2.9.0

# Firebase Authentication
firebase-admin>=6.0.0,<7.0.0
google-auth>=2.0.0

# Environment variables
python-dotenv>=1.0.0

# Web server for production
gunicorn>=20.0.0
